+++ apps/console/src/extensions/i18n/models/extensions.ts
@@ -1701,6 +1701,7 @@ export interface Extensions {
             };
         };
         emailProviders: {
+            configureEmailProvider: string;
             heading: string;
             subHeading: string;
             description: string;
@@ -2332,6 +2333,7 @@ export interface Extensions {
         sidePanel: {
             apiResources: string;
             branding: string;
+            stylesAndText: string;
             monitor: string;
             categories: {
                 apiResources: string;